+++ package.json
@@ -39,6 +39,7 @@
     "postcss-loader": "^6.1.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-icons": "^4.3.1",
     "react-map-to-components": "^1.2.1",
     "regenerator-runtime": "0.13.7",
     "sass-loader": "^10",
